+++ Assets/Scripts/GameObjects/FlyingEnemy.cs
@@ -9,14 +9,13 @@ public class FlyingEnemy : MonoBehaviour, IVisitableEnemy
         health = GetComponent<HealthComponent>();
     }
 
-    public void Accept(IDamageVisitor visitor, int damage)
+    public void Accept(IDamageVisitor visitor, float damage)
     {
         visitor.VisitFlying(this, damage);
     }
 
-    public void ApplyDamage(int damage, Color numberColor)
+    public void ApplyDamage(float damage, Color numberColor)
     {
         health.TakeDamage(damage);
-        //ShowDamageNumber(damage, numberColor);
     }
 }
