+++ Assets/Scripts/Visitors/FireDamageVisitor.cs
@@ -2,13 +2,13 @@ using UnityEngine;
 
 public class FireDamageVisitor : IDamageVisitor
 {
-    public void VisitArmored(ArmoredEnemy enemy, int damage)
+    public void VisitArmored(ArmoredEnemy enemy, float damage)
     {
-        int finalDamage = (int)(damage * 1.5f);
+        float finalDamage = damage * 1.5f;
         enemy.ApplyDamage(finalDamage, new Color(1f, 0.5f, 0f));
     }
 
-    public void VisitFlying(FlyingEnemy enemy, int damage)
+    public void VisitFlying(FlyingEnemy enemy, float damage)
     {
         enemy.ApplyDamage(damage, new Color(1f, 0.5f, 0f));
     }
