+++ Assets/Scripts/Visitors/PhysicalDamageVisitor.cs
@@ -2,13 +2,13 @@ using UnityEngine;
 
 public class PhysicalDamageVisitor: IDamageVisitor
 {
-    public void VisitArmored(ArmoredEnemy enemy, int damage)
+    public void VisitArmored(ArmoredEnemy enemy, float damage)
     {
-        int finalDamage = (int)(damage * 0.5f);
+        float finalDamage = damage * 0.5f;
         enemy.ApplyDamage(finalDamage, Color.red);
     }
 
-    public void VisitFlying(FlyingEnemy enemy, int damage)
+    public void VisitFlying(FlyingEnemy enemy, float damage)
     {
         enemy.ApplyDamage(damage, Color.red);
     }
